remove sub from dcos
