+++ autoarray/fixtures.py
@@ -165,7 +165,6 @@ def make_imaging_7x7_sub_2():
     )
 
 
-
 def make_imaging_covariance_7x7():
     return aa.Imaging(
         data=make_image_7x7(),
@@ -288,11 +287,11 @@ def make_masked_imaging_7x7_no_blur():
 
     return imaging_7x7.apply_mask(mask=make_mask_2d_7x7())
 
+
 def make_masked_imaging_7x7_no_blur_sub_2():
     imaging_7x7 = make_imaging_7x7_no_blur_sub_2()
 
     return imaging_7x7.apply_mask(mask=make_mask_2d_7x7())
-
 
 
 def make_model_image_7x7():
