+++ autoarray/structures/grids/grid_1d_util.py
@@ -22,7 +22,7 @@ def convert_grid_1d(
     This function performs the following and checks and conversions on the input:
 
     1: If the input is a list, convert it to an ndarray.
-    2: Check that the number of sub-pixels in the array is identical to that of the mask.
+    2: Check that the number of pixels in the array is identical to that of the mask.
     3) Map the input ndarray to its `slim` representation.
 
     For a Grid2D, `slim` refers to a 2D NumPy array of shape [total_coordinates, 2] and `native` a 3D NumPy array of
@@ -59,45 +59,38 @@ def convert_grid_1d(
 def grid_1d_slim_via_shape_slim_from(
     shape_slim: Tuple[int],
     pixel_scales: ty.PixelScales,
-    sub_size: int = 1,
     origin: Tuple[float] = (0.0,),
 ) -> np.ndarray:
     """
-    For a sub-grid, every unmasked pixel of its 1D mask with shape (total_pixels,) is divided into a finer uniform
-    grid of shape (total_pixels*sub_size, ). This routine computes the (x) scaled coordinates at the centre of every
-    sub-pixel defined by a 1D shape of the overall grid.
+    This routine computes the (x) scaled coordinates at the centre of every pixel defined by a 1D shape of the
+    overall grid.
 
-    Grid2D are defined from left to right, where the first unmasked sub-pixel corresponds to index 0. Sub-pixels that
-    are part of the same mask array pixel are indexed next to one another, such that the second
-    sub-pixel in the first pixel has index 1, its next sub-pixel has index 2, and so forth.
+    Grid2D are defined from left to right, where the first unmasked pixel corresponds to index 0.
 
-    The sub-grid is returned on an array of shape (total_unmasked_pixels*sub_size, ).
+    The grid is returned on an array of shape (total_unmasked_pixels, ).
 
     Parameters
     ----------
     shape_slim
-        The (x) shape of the 1D array the sub-grid of coordinates is computed for.
+        The (x) shape of the 1D array the grid of coordinates is computed for.
     pixel_scales
         The (x) scaled units to pixel units conversion factor of the 1D mask array.
-    sub_size
-        The size of the sub-grid that each pixel of the 1D mask array is divided into.
     origin
-        The (x) origin of the 1D array, which the sub-grid is shifted around.
+        The (x) origin of the 1D array, which the grid is shifted around.
 
     Returns
     -------
     ndarray
-        A sub grid of (x) scaled coordinates at the centre of every pixel unmasked pixel on the 2D mask
-        array. The sub grid array has dimensions (total_unmasked_pixels*sub_size, ).
+        A grid of (x) scaled coordinates at the centre of every pixel unmasked pixel on the 2D mask
+        array. The grid array has dimensions (total_unmasked_pixels, ).
 
     Examples
     --------
-    sub_grid_1d = grid_1d_via_shape_slim_from(mask=mask, pixel_scales=(0.5, 0.5), sub_size=2, origin=(0.0, 0.0))
+    grid_1d = grid_1d_via_shape_slim_from(mask=mask, pixel_scales=(0.5, 0.5), origin=(0.0, 0.0))
     """
     return grid_1d_slim_via_mask_from(
         mask_1d=np.full(fill_value=False, shape=shape_slim),
         pixel_scales=pixel_scales,
-        sub_size=sub_size,
         origin=origin,
     )
 
@@ -106,72 +99,60 @@ def grid_1d_slim_via_shape_slim_from(
 def grid_1d_slim_via_mask_from(
     mask_1d: np.ndarray,
     pixel_scales: ty.PixelScales,
-    sub_size: int = 1,
     origin: Tuple[float] = (0.0,),
 ) -> np.ndarray:
     """
-    For a sub-grid, every unmasked pixel of its 1D mask with shape (total_pixels,) is divided into a finer uniform
-    grid of shape (total_pixels*sub_size, ). This routine computes the (x) scaled coordinates at the centre of every
-    sub-pixel defined by this 1D mask array.
+    For a grid, every unmasked pixel of its 1D mask with shape (total_pixels,) is divided into a finer uniform
+    grid of shape (total_pixels, ). This routine computes the (x) scaled coordinates at the centre of every
+    pixel defined by this 1D mask array.
 
-    Grid2D are defined from left to right, where the first unmasked sub-pixel corresponds to index 0. Sub-pixels that
-    are part of the same mask array pixel are indexed next to one another, such that the second
-    sub-pixel in the first pixel has index 1, its next sub-pixel has index 2, and so forth.
+    Grid2D are defined from left to right, where the first unmasked pixel corresponds to index 0.
 
-    The sub-grid is returned on an array of shape (total_unmasked_pixels*sub_size, ).
+    The grid is returned on an array of shape (total_unmasked_pixels, ).
 
     Parameters
     ----------
     mask_1d
         A 1D array of bools, where `False` values are unmasked and therefore included as part of the calculated
-        sub-grid.
+        grid.
     pixel_scales
         The (x) scaled units to pixel units conversion factor of the 2D mask array.
-    sub_size
-        The size of the sub-grid that each pixel of the 2D mask array is divided into.
     origin
-        The (x) origin of the 2D array, which the sub-grid is shifted around.
+        The (x) origin of the 2D array, which the grid is shifted around.
 
     Returns
     -------
     ndarray
-        A sub grid of (x) scaled coordinates at the centre of every pixel unmasked pixel on the 2D mask
-        array. The sub grid array has dimensions (total_unmasked_pixels*sub_size, ).
+        A grid of (x) scaled coordinates at the centre of every pixel unmasked pixel on the 2D mask
+        array. The grid array has dimensions (total_unmasked_pixels, ).
 
     Examples
     --------
     mask = np.array([True, False, True, False, False, False])
-    grid_slim = grid_1d_via_mask_from(mask_1d=mask_1d, pixel_scales=(0.5, 0.5), sub_size=1, origin=(0.0, 0.0))
+    grid_slim = grid_1d_via_mask_from(mask_1d=mask_1d, pixel_scales=(0.5, 0.5), origin=(0.0, 0.0))
     """
 
-    total_sub_pixels = mask_1d_util.total_sub_pixels_1d_from(mask_1d, sub_size)
+    total_pixels = mask_1d_util.total_pixels_1d_from(mask_1d)
 
-    grid_1d = np.zeros(shape=(total_sub_pixels,))
+    grid_1d = np.zeros(shape=(total_pixels,))
 
     centres_scaled = geometry_util.central_scaled_coordinate_1d_from(
         shape_slim=mask_1d.shape, pixel_scales=pixel_scales, origin=origin
     )
 
-    sub_index = 0
-
-    x_sub_half = pixel_scales[0] / 2
-    x_sub_step = pixel_scales[0] / (sub_size)
+    index = 0
 
     for x in range(mask_1d.shape[0]):
         if not mask_1d[x]:
-            x_scaled = (x - centres_scaled[0]) * pixel_scales[0]
-
-            for x1 in range(sub_size):
-                grid_1d[sub_index] = (
-                    x_scaled - x_sub_half + x1 * x_sub_step + (x_sub_step / 2.0)
-                )
-                sub_index += 1
+            grid_1d[index] = (x - centres_scaled[0]) * pixel_scales[0]
+            index += 1
 
     return grid_1d
 
 
 def grid_1d_slim_from(
-    grid_1d_native: np.ndarray, mask_1d: np.ndarray, sub_size: int = 1
+    grid_1d_native: np.ndarray,
+    mask_1d: np.ndarray,
 ) -> np.ndarray:
     """
     For a native 1D grid and mask of shape [total_pixels], map the values of all unmasked pixels to a slimmed grid of
@@ -190,8 +171,6 @@ def grid_1d_slim_from(
         The native grid of (x) values which are mapped to the slimmed grid.
     mask_1d
         A 1D array of bools, where `False` values mean unmasked and are included in the mapping.
-    sub_size
-        The size (sub_size x sub_size) of each unmasked pixels sub-array.
 
     Returns
     -------
@@ -202,16 +181,16 @@ def grid_1d_slim_from(
     return array_1d_util.array_1d_slim_from(
         array_1d_native=np.array(grid_1d_native),
         mask_1d=mask_1d,
-        sub_size=sub_size,
     )
 
 
 def grid_1d_native_from(
-    grid_1d_slim: np.ndarray, mask_1d: np.ndarray, sub_size: int = 1
+    grid_1d_slim: np.ndarray,
+    mask_1d: np.ndarray,
 ) -> np.ndarray:
     """
-    For a slimmed 1D grid of shape [total_unmasked_pixels*sub_size], that was computed by extracting the unmasked values
-    from a native 1D grid of shape [total_pixels*sub_size], map the slimmed grid's coordinates back to the native 1D
+    For a slimmed 1D grid of shape [total_unmasked_pixels], that was computed by extracting the unmasked values
+    from a native 1D grid of shape [total_pixels], map the slimmed grid's coordinates back to the native 1D
     grid where masked values are set to zero.
 
     This uses a 1D array 'slim_to_native' where each index gives the 1D pixel indexes of the grid's native unmasked
@@ -227,15 +206,14 @@ def grid_1d_native_from(
         The (x) values of the slimmed 1D grid which are mapped to the native 1D grid.
     mask_1d
         A 1D array of bools, where `False` values means unmasked and are included in the mapping.
-    sub_size
-        The size of each unmasked pixels sub-grid.
 
     Returns
     -------
     ndarray
-        A NumPy array of shape [total_pixels*sub_size] corresponding to the (x) values of the native 1D
+        A NumPy array of shape [total_pixels] corresponding to the (x) values of the native 1D
         mapped from the slimmed grid.
     """
     return array_1d_util.array_1d_native_from(
-        array_1d_slim=grid_1d_slim, mask_1d=mask_1d, sub_size=sub_size
+        array_1d_slim=grid_1d_slim,
+        mask_1d=mask_1d,
     )
