+++ test_autoarray/fit/test_fit_interferometer.py
@@ -6,7 +6,7 @@ import autoarray as aa
 
 def test__visibilities_and_model_are_identical__no_masking__check_values_are_correct():
     real_space_mask = aa.Mask2D(
-        mask=[[False, False], [False, False]], sub_size=1, pixel_scales=(1.0, 1.0)
+        mask=[[False, False], [False, False]], pixel_scales=(1.0, 1.0)
     )
 
     data = aa.Visibilities(visibilities=[1.0 + 2.0j, 3.0 + 4.0j])
@@ -51,7 +51,7 @@ def test__visibilities_and_model_are_identical__no_masking__check_values_are_cor
 
 def test__visibilities_and_model_are_different__no_masking__check_values_are_correct():
     real_space_mask = aa.Mask2D(
-        mask=[[False, False], [False, False]], sub_size=1, pixel_scales=(1.0, 1.0)
+        mask=[[False, False], [False, False]], pixel_scales=(1.0, 1.0)
     )
 
     data = aa.Visibilities(visibilities=[1.0 + 2.0j, 3.0 + 4.0j])
@@ -96,7 +96,7 @@ def test__visibilities_and_model_are_different__no_masking__check_values_are_cor
 
 def test__visibilities_and_model_are_identical__inversion_included__changes_certain_properties():
     real_space_mask = aa.Mask2D(
-        mask=[[False, False], [False, False]], sub_size=1, pixel_scales=(1.0, 1.0)
+        mask=[[False, False], [False, False]], pixel_scales=(1.0, 1.0)
     )
 
     data = aa.Visibilities(visibilities=[1.0 + 2.0j, 3.0 + 4.0j])
