+++ test_autoarray/inversion/inversion/imaging/test_imaging.py
@@ -123,7 +123,6 @@ def test__w_tilde_checks_noise_map_and_raises_exception_if_preloads_dont_match_n
     mask = aa.Mask2D(
         mask=np.array([[True, True, True], [False, False, False], [True, True, True]]),
         pixel_scales=1.0,
-        sub_size=1,
     )
 
     grid = aa.Grid2D.from_mask(mask=mask)
