+++ test_autoarray/mask/test_mask_2d_util.py
@@ -13,14 +13,6 @@ def test__total_pixels_2d_from():
     assert util.mask_2d.total_pixels_2d_from(mask_2d=mask_2d) == 5
 
 
-def test__total_sub_pixels_2d_from():
-    mask_2d = np.array(
-        [[True, False, True], [False, False, False], [True, False, True]]
-    )
-
-    assert util.mask_2d.total_sub_pixels_2d_from(mask_2d, sub_size=2) == 20
-
-
 def test__total_edge_pixels_from_mask():
     mask_2d = np.array(
         [
@@ -742,35 +734,6 @@ def test__mask_2d_elliptical_annular_from__include_centre():
     ).all()
 
 
-def test__oversample_mask_from():
-    mask = np.array(
-        [
-            [True, True, True, True],
-            [True, False, False, True],
-            [True, False, False, True],
-            [True, True, True, True],
-        ]
-    )
-
-    oversample_mask = util.mask_2d.oversample_mask_2d_from(mask=mask, sub_size=2)
-
-    assert (
-        oversample_mask
-        == np.array(
-            [
-                [True, True, True, True, True, True, True, True],
-                [True, True, True, True, True, True, True, True],
-                [True, True, False, False, False, False, True, True],
-                [True, True, False, False, False, False, True, True],
-                [True, True, False, False, False, False, True, True],
-                [True, True, False, False, False, False, True, True],
-                [True, True, True, True, True, True, True, True],
-                [True, True, True, True, True, True, True, True],
-            ]
-        )
-    ).all()
-
-
 def test__mask_2d_via_pixel_coordinates_from():
     mask = util.mask_2d.mask_2d_via_pixel_coordinates_from(
         shape_native=(2, 3), pixel_coordinates=[[0, 1], [1, 1], [1, 2]]
@@ -1258,396 +1221,11 @@ def test__border_slim_indexes_from():
     ).all()
 
 
-def test__sub_border_pixel_slim_indexes_from():
-    mask = np.array(
-        [
-            [True, True, True, True, True, True, True],
-            [True, True, True, True, True, True, True],
-            [True, True, True, True, True, True, True],
-            [True, True, True, True, False, True, True],
-            [True, True, True, True, True, True, True],
-            [True, True, True, True, True, True, True],
-            [True, True, True, True, True, True, True],
-        ]
-    )
-
-    sub_border_pixels = util.mask_2d.sub_border_pixel_slim_indexes_from(
-        mask_2d=mask, sub_size=1
-    )
-
-    assert (sub_border_pixels == np.array([0])).all()
-
-    mask = np.array(
-        [
-            [True, True, True, True, True, True, True],
-            [True, True, True, True, True, True, True],
-            [True, False, False, True, True, True, True],
-            [True, True, True, True, False, True, True],
-            [True, True, True, True, True, True, True],
-            [True, True, True, True, True, True, True],
-            [True, True, True, True, True, True, True],
-        ]
-    )
-
-    sub_border_pixels = util.mask_2d.sub_border_pixel_slim_indexes_from(
-        mask_2d=mask, sub_size=1
-    )
-
-    assert (sub_border_pixels == np.array([0, 1, 2])).all()
-
-    mask = np.array(
-        [
-            [True, True, True, True, True, True, True],
-            [True, True, True, True, True, True, True],
-            [True, True, False, False, False, True, True],
-            [True, True, False, False, False, True, True],
-            [True, True, False, False, False, True, True],
-            [True, True, True, True, True, True, True],
-            [True, True, True, True, True, True, True],
-        ]
-    )
-
-    sub_border_pixels = util.mask_2d.sub_border_pixel_slim_indexes_from(
-        mask_2d=mask, sub_size=1
-    )
-
-    assert (sub_border_pixels == np.array([0, 1, 2, 3, 5, 6, 7, 8])).all()
-
-    mask = np.array(
-        [
-            [True, True, True, True, True, True, True],
-            [True, True, True, True, True, True, True],
-            [True, True, True, True, True, True, True],
-            [True, True, True, True, False, True, True],
-            [True, True, True, True, True, True, True],
-            [True, True, True, True, True, True, True],
-            [True, True, True, True, True, True, True],
-        ]
-    )
-
-    sub_border_pixels = util.mask_2d.sub_border_pixel_slim_indexes_from(
-        mask_2d=mask, sub_size=2
-    )
-
-    assert (sub_border_pixels == np.array([3])).all()
-
-    mask = np.array(
-        [
-            [True, True, True, True, True, True, True],
-            [True, True, True, True, True, True, True],
-            [True, True, False, False, False, True, True],
-            [True, True, False, False, False, True, True],
-            [True, True, False, False, False, True, True],
-            [True, True, True, True, True, True, True],
-            [True, True, True, True, True, True, True],
-        ]
-    )
-
-    sub_border_pixels = util.mask_2d.sub_border_pixel_slim_indexes_from(
-        mask_2d=mask, sub_size=2
-    )
-
-    assert (sub_border_pixels == np.array([0, 5, 9, 14, 23, 26, 31, 35])).all()
-
-    mask = np.array(
-        [
-            [True, True, True, True, True, True, True],
-            [True, True, True, True, True, True, True],
-            [True, True, False, False, False, True, True],
-            [True, True, False, False, False, True, True],
-            [True, True, False, False, False, True, True],
-            [True, True, True, True, True, True, True],
-            [True, True, True, True, True, True, True],
-        ]
-    )
-
-    sub_border_pixels = util.mask_2d.sub_border_pixel_slim_indexes_from(
-        mask_2d=mask, sub_size=3
-    )
-
-    assert (sub_border_pixels == np.array([0, 11, 20, 33, 53, 60, 71, 80])).all()
-
-    mask = np.array(
-        [
-            [True, True, True, True, True, True, True, True, True],
-            [True, False, False, False, False, False, False, False, True],
-            [True, False, True, True, True, True, True, False, True],
-            [True, False, True, False, False, False, True, False, True],
-            [True, False, True, False, True, False, True, False, True],
-            [True, False, True, False, False, False, True, False, True],
-            [True, False, True, True, True, True, True, False, True],
-            [True, False, False, False, False, False, False, False, True],
-            [True, True, True, True, True, True, True, True, True],
-            [True, True, True, True, True, True, True, True, True],
-        ]
-    )
-
-    sub_border_pixels = util.mask_2d.sub_border_pixel_slim_indexes_from(
-        mask_2d=mask, sub_size=2
-    )
-
-    assert (
-        sub_border_pixels
-        == np.array(
-            [
-                0,
-                4,
-                8,
-                13,
-                17,
-                21,
-                25,
-                28,
-                33,
-                36,
-                53,
-                58,
-                71,
-                74,
-                91,
-                94,
-                99,
-                102,
-                106,
-                110,
-                115,
-                119,
-                123,
-                127,
-            ]
-        )
-    ).all()
-
-    mask = np.array(
-        [
-            [True, True, True, True, True, True, True, True, True, True],
-            [True, False, False, False, False, False, False, False, True, True],
-            [True, False, True, True, True, True, True, False, True, True],
-            [True, False, True, False, False, False, True, False, True, True],
-            [True, False, True, False, True, False, True, False, True, True],
-            [True, False, True, False, False, False, True, False, True, True],
-            [True, False, True, True, True, True, True, False, True, True],
-            [True, False, False, False, False, False, False, False, True, True],
-            [True, True, True, True, True, True, True, True, True, True],
-        ]
-    )
-
-    sub_border_pixels = util.mask_2d.sub_border_pixel_slim_indexes_from(
-        mask_2d=mask, sub_size=2
-    )
-
-    assert (
-        sub_border_pixels
-        == np.array(
-            [
-                0,
-                4,
-                8,
-                13,
-                17,
-                21,
-                25,
-                28,
-                33,
-                36,
-                53,
-                58,
-                71,
-                74,
-                91,
-                94,
-                99,
-                102,
-                106,
-                110,
-                115,
-                119,
-                123,
-                127,
-            ]
-        )
-    ).all()
-
-
-def test__slim_index_for_sub_slim_index_via_mask_2d_from():
-    mask = np.array([[True, True, True], [True, False, True], [True, True, True]])
-
-    slim_index_for_sub_slim_index = (
-        util.mask_2d.slim_index_for_sub_slim_index_via_mask_2d_from(mask, sub_size=2)
-    )
-
-    assert (slim_index_for_sub_slim_index == np.array([0, 0, 0, 0])).all()
-
-    mask = np.array([[True, True, True], [False, False, False], [True, True, True]])
-
-    slim_index_for_sub_slim_index = (
-        util.mask_2d.slim_index_for_sub_slim_index_via_mask_2d_from(mask, sub_size=2)
-    )
-
-    assert (
-        slim_index_for_sub_slim_index == np.array([0, 0, 0, 0, 1, 1, 1, 1, 2, 2, 2, 2])
-    ).all()
-
-    mask = np.array([[True, True, True], [False, False, False], [True, True, True]])
-
-    slim_index_for_sub_slim_index = (
-        util.mask_2d.slim_index_for_sub_slim_index_via_mask_2d_from(mask, sub_size=3)
-    )
-
-    assert (
-        slim_index_for_sub_slim_index
-        == np.array(
-            [
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                1,
-                1,
-                1,
-                1,
-                1,
-                1,
-                1,
-                1,
-                1,
-                2,
-                2,
-                2,
-                2,
-                2,
-                2,
-                2,
-                2,
-                2,
-            ]
-        )
-    ).all()
-
-
-def test__sub_slim_indexes_for_slim_index_via_mask_2d_from():
-    mask = np.array([[True, True, True], [True, False, True], [True, True, True]])
-
-    sub_mask_1d_indexes_for_mask_1d_index = (
-        util.mask_2d.sub_slim_indexes_for_slim_index_via_mask_2d_from(mask, sub_size=2)
-    )
-
-    assert sub_mask_1d_indexes_for_mask_1d_index == [[0, 1, 2, 3]]
-
-    mask = np.array([[True, True, True], [False, False, False], [True, True, True]])
-
-    sub_mask_1d_indexes_for_mask_1d_index = (
-        util.mask_2d.sub_slim_indexes_for_slim_index_via_mask_2d_from(mask, sub_size=2)
-    )
-
-    assert sub_mask_1d_indexes_for_mask_1d_index == [
-        [0, 1, 2, 3],
-        [4, 5, 6, 7],
-        [8, 9, 10, 11],
-    ]
-
-    mask = np.array([[True, True, True], [False, False, False], [True, True, True]])
-
-    sub_mask_1d_indexes_for_mask_1d_index = (
-        util.mask_2d.sub_slim_indexes_for_slim_index_via_mask_2d_from(mask, sub_size=3)
-    )
-
-    assert sub_mask_1d_indexes_for_mask_1d_index == [
-        [0, 1, 2, 3, 4, 5, 6, 7, 8],
-        [9, 10, 11, 12, 13, 14, 15, 16, 17],
-        [18, 19, 20, 21, 22, 23, 24, 25, 26],
-    ]
-
-
-def test__sub_slim_index_for_sub_native_index_from():
-    mask = np.full(fill_value=False, shape=(3, 3))
-
-    sub_mask_1d_index_for_sub_mask_index = (
-        util.mask_2d.sub_slim_index_for_sub_native_index_from(sub_mask_2d=mask)
-    )
-
-    assert (
-        sub_mask_1d_index_for_sub_mask_index
-        == np.array([[0, 1, 2], [3, 4, 5], [6, 7, 8]])
-    ).all()
-
-    mask = np.full(fill_value=False, shape=(2, 3))
-
-    sub_mask_1d_index_for_sub_mask_index = (
-        util.mask_2d.sub_slim_index_for_sub_native_index_from(sub_mask_2d=mask)
-    )
-
-    assert (
-        sub_mask_1d_index_for_sub_mask_index == np.array([[0, 1, 2], [3, 4, 5]])
-    ).all()
-
-    mask = np.full(fill_value=False, shape=(3, 2))
-
-    sub_mask_1d_index_for_sub_mask_index = (
-        util.mask_2d.sub_slim_index_for_sub_native_index_from(sub_mask_2d=mask)
-    )
-
-    assert (
-        sub_mask_1d_index_for_sub_mask_index == np.array([[0, 1], [2, 3], [4, 5]])
-    ).all()
-
-    mask = np.array([[False, True, False], [True, True, False], [False, False, True]])
-
-    sub_mask_1d_index_for_sub_mask_index = (
-        util.mask_2d.sub_slim_index_for_sub_native_index_from(sub_mask_2d=mask)
-    )
-
-    assert (
-        sub_mask_1d_index_for_sub_mask_index
-        == np.array([[0, -1, 1], [-1, -1, 2], [3, 4, -1]])
-    ).all()
-
-    mask = np.array(
-        [
-            [False, True, True, False],
-            [True, True, False, False],
-            [False, False, True, False],
-        ]
-    )
-
-    sub_mask_1d_index_for_sub_mask_index = (
-        util.mask_2d.sub_slim_index_for_sub_native_index_from(sub_mask_2d=mask)
-    )
-
-    assert (
-        sub_mask_1d_index_for_sub_mask_index
-        == np.array([[0, -1, -1, 1], [-1, -1, 2, 3], [4, 5, -1, 6]])
-    ).all()
-
-    mask = np.array(
-        [
-            [False, True, False],
-            [True, True, False],
-            [False, False, True],
-            [False, False, True],
-        ]
-    )
-
-    sub_mask_1d_index_for_sub_mask_index = (
-        util.mask_2d.sub_slim_index_for_sub_native_index_from(sub_mask_2d=mask)
-    )
-
-    assert (
-        sub_mask_1d_index_for_sub_mask_index
-        == np.array([[0, -1, 1], [-1, -1, 2], [3, 4, -1], [5, 6, -1]])
-    ).all()
-
-
 def test__native_index_for_slim_index_2d_from():
     mask = np.array([[True, True, True], [True, False, True], [True, True, True]])
 
     sub_mask_index_for_sub_mask_1d_index = (
-        util.mask_2d.native_index_for_slim_index_2d_from(mask_2d=mask, sub_size=1)
+        util.mask_2d.native_index_for_slim_index_2d_from(mask_2d=mask)
     )
 
     assert (sub_mask_index_for_sub_mask_1d_index == np.array([[1, 1]])).all()
@@ -1662,75 +1240,12 @@ def test__native_index_for_slim_index_2d_from():
     )
 
     sub_mask_index_for_sub_mask_1d_index = (
-        util.mask_2d.native_index_for_slim_index_2d_from(mask_2d=mask, sub_size=1)
+        util.mask_2d.native_index_for_slim_index_2d_from(mask_2d=mask)
     )
 
     assert (
         sub_mask_index_for_sub_mask_1d_index
         == np.array([[0, 1], [1, 0], [1, 1], [1, 2], [2, 1], [3, 2]])
-    ).all()
-
-    mask = np.array([[True, True, True], [True, False, True], [True, True, True]])
-
-    sub_mask_index_for_sub_mask_1d_index = (
-        util.mask_2d.native_index_for_slim_index_2d_from(mask_2d=mask, sub_size=2)
-    )
-
-    assert (
-        sub_mask_index_for_sub_mask_1d_index
-        == np.array([[2, 2], [2, 3], [3, 2], [3, 3]])
-    ).all()
-
-    mask = np.array([[True, False, True], [False, False, False], [True, False, True]])
-
-    sub_mask_index_for_sub_mask_1d_index = (
-        util.mask_2d.native_index_for_slim_index_2d_from(mask_2d=mask, sub_size=2)
-    )
-
-    assert (
-        sub_mask_index_for_sub_mask_1d_index
-        == np.array(
-            [
-                [0, 2],
-                [0, 3],
-                [1, 2],
-                [1, 3],
-                [2, 0],
-                [2, 1],
-                [3, 0],
-                [3, 1],
-                [2, 2],
-                [2, 3],
-                [3, 2],
-                [3, 3],
-                [2, 4],
-                [2, 5],
-                [3, 4],
-                [3, 5],
-                [4, 2],
-                [4, 3],
-                [5, 2],
-                [5, 3],
-            ]
-        )
-    ).all()
-
-    mask = np.array(
-        [
-            [True, True, True],
-            [True, False, True],
-            [True, True, True],
-            [True, True, False],
-        ]
-    )
-
-    sub_mask_index_for_sub_mask_1d_index = (
-        util.mask_2d.native_index_for_slim_index_2d_from(mask_2d=mask, sub_size=2)
-    )
-
-    assert (
-        sub_mask_index_for_sub_mask_1d_index
-        == np.array([[2, 2], [2, 3], [3, 2], [3, 3], [6, 4], [6, 5], [7, 4], [7, 5]])
     ).all()
 
 
@@ -1892,59 +1407,3 @@ def test__buffed_mask_2d_from():
             ]
         )
     ).all()
-
-
-def test__mask_2d_neighbors_from():
-    mask = np.array(
-        [
-            [True, True, True, True],
-            [True, False, False, True],
-            [True, False, False, True],
-            [True, True, True, True],
-        ]
-    )
-
-    mask_neighbors = util.mask_2d.mask_2d_neighbors_from(mask_2d=mask)
-
-    assert (mask_neighbors == np.array([1, 3, 3, 2])).all()
-
-    mask = np.array(
-        [
-            [True, True, True, True],
-            [True, False, False, True],
-            [True, False, False, True],
-            [True, False, False, True],
-            [True, True, True, True],
-        ]
-    )
-
-    mask_neighbors = util.mask_2d.mask_2d_neighbors_from(mask_2d=mask)
-
-    assert (mask_neighbors == np.array([1, 3, 3, 5, 5, 4])).all()
-
-    mask = np.array([[False, False], [False, False]])
-
-    mask_neighbors = util.mask_2d.mask_2d_neighbors_from(mask_2d=mask)
-
-    assert (mask_neighbors == np.array([1, 3, 3, 2])).all()
-
-    mask = np.array([[False, False], [False, False], [False, False]])
-
-    mask_neighbors = util.mask_2d.mask_2d_neighbors_from(mask_2d=mask)
-
-    assert (mask_neighbors == np.array([1, 3, 3, 5, 5, 4])).all()
-
-
-def test__mask_2d_neighbors_from__pixel_with_no_adjacent_neighbor__gives_minus_1():
-    mask = np.array(
-        [
-            [True, True, True, True],
-            [True, False, True, False],
-            [True, False, False, True],
-            [True, True, True, True],
-        ]
-    )
-
-    mask_neighbors = util.mask_2d.mask_2d_neighbors_from(mask_2d=mask)
-
-    assert (mask_neighbors == np.array([2, -1, 3, 2])).all()
