+++ test_autoarray/structures/arrays/test_array_1d_util.py
@@ -16,7 +16,8 @@ def test__array_1d_slim_from():
     array_1d_native = np.array([1.0, 2.0, 3.0, 4.0])
 
     array_1d_slim = aa.util.array_1d.array_1d_slim_from(
-        array_1d_native=array_1d_native, mask_1d=mask_1d, sub_size=1
+        array_1d_native=array_1d_native,
+        mask_1d=mask_1d,
     )
 
     assert (array_1d_slim == np.array([1.0, 2.0, 3.0, 4.0])).all()
@@ -26,22 +27,11 @@ def test__array_1d_slim_from():
     array_1d_native = np.array([1.0, 2.0, 3.0, 4.0, 5.0, 6.0])
 
     array_1d_slim = aa.util.array_1d.array_1d_slim_from(
-        array_1d_native=array_1d_native, mask_1d=mask_1d, sub_size=1
+        array_1d_native=array_1d_native,
+        mask_1d=mask_1d,
     )
 
     assert (array_1d_slim == np.array([2.0, 3.0, 5.0, 6.0])).all()
-
-    mask_1d = np.array([True, True, False, True, False, False])
-
-    array_1d_native = np.array(
-        [1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0]
-    )
-
-    array_1d_slim = aa.util.array_1d.array_1d_slim_from(
-        array_1d_native=array_1d_native, mask_1d=mask_1d, sub_size=2
-    )
-
-    assert (array_1d_slim == np.array([5.0, 6.0, 9.0, 10.0, 11.0, 12.0])).all()
 
 
 def test__array_1d_native_from():
@@ -50,7 +40,8 @@ def test__array_1d_native_from():
     array_1d_slim = np.array([1.0, 2.0, 3.0, 4.0])
 
     array_1d_native = aa.util.array_1d.array_1d_native_from(
-        array_1d_slim=array_1d_slim, mask_1d=mask_1d, sub_size=1
+        array_1d_slim=array_1d_slim,
+        mask_1d=mask_1d,
     )
 
     assert (array_1d_native == np.array([1.0, 2.0, 3.0, 4.0])).all()
@@ -60,23 +51,11 @@ def test__array_1d_native_from():
     array_1d_slim = np.array([1.0, 2.0, 3.0, 4.0])
 
     array_1d_native = aa.util.array_1d.array_1d_native_from(
-        array_1d_slim=array_1d_slim, mask_1d=mask_1d, sub_size=1
+        array_1d_slim=array_1d_slim,
+        mask_1d=mask_1d,
     )
 
     assert (array_1d_native == np.array([1.0, 2.0, 0.0, 0.0, 3.0, 4.0])).all()
-
-    mask_1d = np.array([True, False, False, True, False, False])
-
-    array_1d_slim = np.array([1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0])
-
-    array_1d_native = aa.util.array_1d.array_1d_native_from(
-        array_1d_slim=array_1d_slim, mask_1d=mask_1d, sub_size=2
-    )
-
-    assert (
-        array_1d_native
-        == np.array([0.0, 0.0, 1.0, 2.0, 3.0, 4.0, 0.0, 0.0, 5.0, 6.0, 7.0, 8.0])
-    ).all()
 
 
 def test__numpy_array_1d_via_fits_from():
