+++ test_autoarray/structures/arrays/test_array_2d_util.py
@@ -517,13 +517,14 @@ def test__index_slim_for_index_2d_from():
     assert (index_slim_for_index_2d == np.array([1, 4, 7, 8, 0, 2, 3, 5, 6])).all()
 
 
-def test__array_2d_slim_from__sub_size_1():
+def test__array_2d_slim_from():
     array_2d = np.array([[1, 2, 3], [4, 5, 6], [7, 8, 9]])
 
     mask = np.array([[True, True, True], [True, False, True], [True, True, True]])
 
     array_2d_slim = util.array_2d.array_2d_slim_from(
-        mask_2d=mask, array_2d_native=array_2d, sub_size=1
+        mask_2d=mask,
+        array_2d_native=array_2d,
     )
 
     assert (array_2d_slim == np.array([5])).all()
@@ -533,7 +534,8 @@ def test__array_2d_slim_from__sub_size_1():
     mask = np.array([[True, False, True], [False, False, False], [True, False, True]])
 
     array_2d_slim = util.array_2d.array_2d_slim_from(
-        mask_2d=mask, array_2d_native=array_2d, sub_size=1
+        mask_2d=mask,
+        array_2d_native=array_2d,
     )
 
     assert (array_2d_slim == np.array([2, 4, 5, 6, 8])).all()
@@ -549,7 +551,8 @@ def test__array_2d_slim_from__sub_size_1():
     )
 
     array_2d_slim = util.array_2d.array_2d_slim_from(
-        mask_2d=mask, array_2d_native=array_2d, sub_size=1
+        mask_2d=mask,
+        array_2d_native=array_2d,
     )
 
     assert (array_2d_slim == np.array([2, 5, 6, 7, 10, 12])).all()
@@ -566,149 +569,11 @@ def test__array_2d_slim_from__sub_size_1():
     )
 
     array_2d_slim = util.array_2d.array_2d_slim_from(
-        mask_2d=mask, array_2d_native=array_2d, sub_size=1
+        mask_2d=mask,
+        array_2d_native=array_2d,
     )
 
     assert (array_2d_slim == np.array([2, 4, 5, 6, 8])).all()
-
-
-def test__array_2d_slim_from__sub_size_2_or_3():
-    sub_array_2d = np.array(
-        [
-            [1, 2, 3, 4, 5, 6],
-            [7, 8, 9, 10, 11, 12],
-            [13, 14, 15, 16, 17, 18],
-            [1, 2, 3, 4, 5, 6],
-            [7, 8, 9, 10, 11, 12],
-            [13, 14, 15, 16, 17, 18],
-        ]
-    )
-
-    mask = np.array([[True, True, True], [True, False, True], [True, True, True]])
-
-    sub_array_1d = util.array_2d.array_2d_slim_from(
-        array_2d_native=sub_array_2d, mask_2d=mask, sub_size=2
-    )
-
-    assert (sub_array_1d == np.array([15, 16, 3, 4])).all()
-
-    mask = np.array([[True, False, True], [True, False, True], [True, True, False]])
-
-    sub_array_1d = util.array_2d.array_2d_slim_from(
-        array_2d_native=sub_array_2d, mask_2d=mask, sub_size=2
-    )
-
-    assert (sub_array_1d == np.array([3, 4, 9, 10, 15, 16, 3, 4, 11, 12, 17, 18])).all()
-
-    sub_array_2d = np.array(
-        [
-            [1, 2, 3, 4, 5, 6, 7, 7, 7],
-            [7, 8, 9, 10, 11, 12, 7, 7, 7],
-            [13, 14, 15, 16, 17, 18, 7, 7, 7],
-            [1, 2, 3, 4, 5, 6, 7, 7, 7],
-            [7, 8, 9, 10, 11, 12, 7, 7, 7],
-            [13, 14, 15, 16, 17, 18, 7, 7, 7],
-        ]
-    )
-
-    mask = np.array(
-        [
-            [True, False, True, True],
-            [False, False, False, True],
-            [True, False, True, False],
-        ]
-    )
-
-    sub_array_1d = util.array_2d.array_2d_slim_from(
-        array_2d_native=sub_array_2d, mask_2d=mask, sub_size=2
-    )
-
-    assert (
-        sub_array_1d
-        == np.array(
-            [
-                3,
-                4,
-                9,
-                10,
-                13,
-                14,
-                1,
-                2,
-                15,
-                16,
-                3,
-                4,
-                17,
-                18,
-                5,
-                6,
-                9,
-                10,
-                15,
-                16,
-                7,
-                7,
-                7,
-                7,
-            ]
-        )
-    ).all()
-
-    sub_array_2d = np.array(
-        [
-            [1, 2, 3, 4, 5, 6],
-            [7, 8, 9, 10, 11, 12],
-            [13, 14, 15, 16, 17, 18],
-            [1, 2, 3, 4, 5, 6],
-            [7, 8, 9, 10, 11, 12],
-            [13, 14, 15, 16, 17, 18],
-            [7, 7, 7, 7, 7, 7],
-            [7, 7, 7, 7, 7, 7],
-        ]
-    )
-
-    mask = np.array(
-        [
-            [True, False, True],
-            [False, False, False],
-            [True, False, True],
-            [True, True, True],
-        ]
-    )
-
-    sub_array_1d = util.array_2d.array_2d_slim_from(
-        array_2d_native=sub_array_2d, mask_2d=mask, sub_size=2
-    )
-
-    assert (
-        sub_array_1d
-        == np.array(
-            [3, 4, 9, 10, 13, 14, 1, 2, 15, 16, 3, 4, 17, 18, 5, 6, 9, 10, 15, 16]
-        )
-    ).all()
-
-    sub_array_2d = np.array(
-        [
-            [1, 2, 3, 4, 5, 6],
-            [7, 8, 9, 10, 11, 12],
-            [13, 14, 15, 16, 17, 18],
-            [1, 2, 3, 4, 5, 6],
-            [7, 8, 9, 10, 11, 12],
-            [13, 14, 15, 16, 17, 18],
-        ]
-    )
-
-    mask = np.array([[False, True], [True, False]])
-
-    sub_array_1d = util.array_2d.array_2d_slim_from(
-        array_2d_native=sub_array_2d, mask_2d=mask, sub_size=3
-    )
-
-    assert (
-        sub_array_1d
-        == np.array([1, 2, 3, 7, 8, 9, 13, 14, 15, 4, 5, 6, 10, 11, 12, 16, 17, 18])
-    ).all()
 
 
 def test__array_2d_slim_from__complex_array():
@@ -723,19 +588,21 @@ def test__array_2d_slim_from__complex_array():
     mask = np.array([[True, True, True], [True, False, True], [True, True, True]])
 
     array_2d_slim = util.array_2d.array_2d_slim_complex_from(
-        mask=mask, array_2d_native=array_2d, sub_size=1
+        mask=mask,
+        array_2d_native=array_2d,
     )
 
     assert (array_2d_slim == np.array([5 + 5j])).all()
 
 
-def test__array_2d_native_from__sub_size_1():
+def test__array_2d_native_from():
     array_2d_slim = np.array([1.0, 2.0, 3.0, 4.0])
 
     mask = np.full(fill_value=False, shape=(2, 2))
 
     array_2d = util.array_2d.array_2d_native_from(
-        array_2d_slim=array_2d_slim, mask_2d=mask, sub_size=1
+        array_2d_slim=array_2d_slim,
+        mask_2d=mask,
     )
 
     assert (array_2d == np.array([[1.0, 2.0], [3.0, 4.0]])).all()
@@ -745,7 +612,8 @@ def test__array_2d_native_from__sub_size_1():
     mask = np.array([[False, False], [False, True]])
 
     array_2d = util.array_2d.array_2d_native_from(
-        array_2d_slim=array_2d_slim, mask_2d=mask, sub_size=1
+        array_2d_slim=array_2d_slim,
+        mask_2d=mask,
     )
 
     assert (array_2d == np.array([[1.0, 2.0], [3.0, 0.0]])).all()
@@ -761,37 +629,14 @@ def test__array_2d_native_from__sub_size_1():
     )
 
     array_2d = util.array_2d.array_2d_native_from(
-        array_2d_slim=array_2d_slim, mask_2d=mask, sub_size=1
+        array_2d_slim=array_2d_slim,
+        mask_2d=mask,
     )
 
     assert (
         array_2d
         == np.array(
             [[1.0, 2.0, 0.0, 0.0], [3.0, 0.0, 0.0, 0.0], [-1.0, -2.0, 0.0, -3.0]]
-        )
-    ).all()
-
-
-def test__array_2d_native_from__sub_size_2():
-    array_2d_slim = np.array(
-        [1.0, 1.0, 1.0, 1.0, 2.0, 2.0, 2.0, 2.0, 3.0, 3.0, 3.0, 4.0]
-    )
-
-    mask = np.array([[False, False], [False, True]])
-
-    array_2d = util.array_2d.array_2d_native_from(
-        array_2d_slim=array_2d_slim, mask_2d=mask, sub_size=2
-    )
-
-    assert (
-        array_2d
-        == np.array(
-            [
-                [1.0, 1.0, 2.0, 2.0],
-                [1.0, 1.0, 2.0, 2.0],
-                [3.0, 3.0, 0.0, 0.0],
-                [3.0, 4.0, 0.0, 0.0],
-            ]
         )
     ).all()
 
@@ -803,7 +648,7 @@ def test__array_2d_native_from__compelx_array():
 
     array_2d = util.array_2d.array_2d_native_complex_via_indexes_from(
         array_2d_slim=array_2d_slim,
-        sub_shape_native=(2, 2),
+        shape_native=(2, 2),
         native_index_for_slim_index_2d=np.array(
             [[0, 0], [0, 1], [1, 0], [1, 1]], dtype="int"
         ),
