+++ test_autoarray/structures/grids/test_grid_2d_util.py
@@ -7,7 +7,8 @@ def test__grid_2d_slim_via_mask_from():
     mask = np.array([[True, True, True], [True, False, True], [True, True, True]])
 
     grid = aa.util.grid_2d.grid_2d_slim_via_mask_from(
-        mask_2d=mask, pixel_scales=(3.0, 6.0), sub_size=1
+        mask_2d=mask,
+        pixel_scales=(3.0, 6.0),
     )
 
     assert (grid[0] == np.array([0.0, 0.0])).all()
@@ -21,7 +22,8 @@ def test__grid_2d_slim_via_mask_from():
     )
 
     grid = aa.util.grid_2d.grid_2d_slim_via_mask_from(
-        mask_2d=mask, pixel_scales=(3.0, 3.0), sub_size=1
+        mask_2d=mask,
+        pixel_scales=(3.0, 3.0),
     )
 
     assert (
@@ -47,7 +49,7 @@ def test__grid_2d_slim_via_mask_from():
     )
 
     grid = aa.util.grid_2d.grid_2d_slim_via_mask_from(
-        mask_2d=mask, pixel_scales=(3.0, 3.0), sub_size=1, origin=(1.0, 2.0)
+        mask_2d=mask, pixel_scales=(3.0, 3.0), origin=(1.0, 2.0)
     )
 
     assert grid == pytest.approx(
@@ -66,119 +68,13 @@ def test__grid_2d_slim_via_mask_from():
 
     mask = np.array([[True, True, False], [False, False, False], [True, True, False]])
 
-    grid = aa.util.grid_2d.grid_2d_slim_via_mask_from(
-        mask_2d=mask, pixel_scales=(3.0, 3.0), sub_size=2
-    )
-
-    assert (
-        grid
-        == np.array(
-            [
-                [3.75, 2.25],
-                [3.75, 3.75],
-                [2.25, 2.25],
-                [2.25, 3.75],
-                [0.75, -3.75],
-                [0.75, -2.25],
-                [-0.75, -3.75],
-                [-0.75, -2.25],
-                [0.75, -0.75],
-                [0.75, 0.75],
-                [-0.75, -0.75],
-                [-0.75, 0.75],
-                [0.75, 2.25],
-                [0.75, 3.75],
-                [-0.75, 2.25],
-                [-0.75, 3.75],
-                [-2.25, 2.25],
-                [-2.25, 3.75],
-                [-3.75, 2.25],
-                [-3.75, 3.75],
-            ]
-        )
-    ).all()
-
-    mask = np.array([[True, True, True], [True, False, True], [True, True, True]])
-
-    grid = aa.util.grid_2d.grid_2d_slim_via_mask_from(
-        mask_2d=mask, pixel_scales=(3.0, 3.0), sub_size=3
-    )
-
-    assert (
-        grid
-        == np.array(
-            [
-                [
-                    [1.0, -1.0],
-                    [1.0, 0.0],
-                    [1.0, 1.0],
-                    [0.0, -1.0],
-                    [0.0, 0.0],
-                    [0.0, 1.0],
-                    [-1.0, -1.0],
-                    [-1.0, 0.0],
-                    [-1.0, 1.0],
-                ]
-            ]
-        )
-    ).all()
-
-    mask = np.array(
-        [
-            [True, True, True, False],
-            [True, False, False, True],
-            [False, True, False, True],
-        ]
-    )
-
-    grid = aa.util.grid_2d.grid_2d_slim_via_mask_from(
-        mask_2d=mask, pixel_scales=(3.0, 3.0), sub_size=2
-    )
-
-    assert (
-        grid
-        == np.array(
-            [
-                [3.75, 3.75],
-                [3.75, 5.25],
-                [2.25, 3.75],
-                [2.25, 5.25],
-                [0.75, -2.25],
-                [0.75, -0.75],
-                [-0.75, -2.25],
-                [-0.75, -0.75],
-                [0.75, 0.75],
-                [0.75, 2.25],
-                [-0.75, 0.75],
-                [-0.75, 2.25],
-                [-2.25, -5.25],
-                [-2.25, -3.75],
-                [-3.75, -5.25],
-                [-3.75, -3.75],
-                [-2.25, 0.75],
-                [-2.25, 2.25],
-                [-3.75, 0.75],
-                [-3.75, 2.25],
-            ]
-        )
-    ).all()
-
-    mask = np.array([[True, True, True], [True, False, True], [True, True, True]])
-
-    grid = aa.util.grid_2d.grid_2d_slim_via_mask_from(
-        mask_2d=mask, pixel_scales=(3.0, 6.0), sub_size=2, origin=(1.0, 1.0)
-    )
-
-    assert grid[0:4] == pytest.approx(
-        np.array([[1.75, -0.5], [1.75, 2.5], [0.25, -0.5], [0.25, 2.5]]), 1e-4
-    )
-
 
 def test__grid_2d_via_mask_from():
     mask = np.array([[False, True, True], [True, True, False], [True, True, True]])
 
     grid_2d = aa.util.grid_2d.grid_2d_via_mask_from(
-        mask_2d=mask, pixel_scales=(3.0, 6.0), sub_size=1
+        mask_2d=mask,
+        pixel_scales=(3.0, 6.0),
     )
 
     assert (
@@ -192,28 +88,11 @@ def test__grid_2d_via_mask_from():
         )
     ).all()
 
-    mask = np.array([[False, True], [True, False]])
-
-    grid_2d = aa.util.grid_2d.grid_2d_via_mask_from(
-        mask_2d=mask, pixel_scales=(3.0, 6.0), sub_size=2
-    )
-
-    assert (
-        grid_2d
-        == np.array(
-            [
-                [[2.25, -4.5], [2.25, -1.5], [0.0, 0.0], [0.0, 0.0]],
-                [[0.75, -4.5], [0.75, -1.5], [0.0, 0.0], [0.0, 0.0]],
-                [[0.0, 0.0], [0.0, 0.0], [-0.75, 1.5], [-0.75, 4.5]],
-                [[0.0, 0.0], [0.0, 0.0], [-2.25, 1.5], [-2.25, 4.5]],
-            ]
-        )
-    ).all()
-
 
 def test__grid_2d_slim_via_shape_native_from():
     grid_2d = aa.util.grid_2d.grid_2d_slim_via_shape_native_from(
-        shape_native=(2, 3), pixel_scales=(1.0, 1.0), sub_size=1
+        shape_native=(2, 3),
+        pixel_scales=(1.0, 1.0),
     )
 
     assert (
@@ -231,7 +110,8 @@ def test__grid_2d_slim_via_shape_native_from():
     ).all()
 
     grid_2d = aa.util.grid_2d.grid_2d_slim_via_shape_native_from(
-        shape_native=(3, 2), pixel_scales=(1.0, 1.0), sub_size=1
+        shape_native=(3, 2),
+        pixel_scales=(1.0, 1.0),
     )
 
     assert (
@@ -249,7 +129,7 @@ def test__grid_2d_slim_via_shape_native_from():
     ).all()
 
     grid_2d = aa.util.grid_2d.grid_2d_slim_via_shape_native_from(
-        shape_native=(3, 2), pixel_scales=(1.0, 1.0), sub_size=1, origin=(3.0, -2.0)
+        shape_native=(3, 2), pixel_scales=(1.0, 1.0), origin=(3.0, -2.0)
     )
 
     assert (
@@ -266,71 +146,11 @@ def test__grid_2d_slim_via_shape_native_from():
         )
     ).all()
 
-    grid = aa.util.grid_2d.grid_2d_slim_via_shape_native_from(
-        shape_native=(3, 3), pixel_scales=(1.0, 1.0), sub_size=2
-    )
-
-    assert (
-        grid
-        == np.array(
-            [
-                [1.25, -1.25],
-                [1.25, -0.75],
-                [0.75, -1.25],
-                [0.75, -0.75],
-                [1.25, -0.25],
-                [1.25, 0.25],
-                [0.75, -0.25],
-                [0.75, 0.25],
-                [1.25, 0.75],
-                [1.25, 1.25],
-                [0.75, 0.75],
-                [0.75, 1.25],
-                [0.25, -1.25],
-                [0.25, -0.75],
-                [-0.25, -1.25],
-                [-0.25, -0.75],
-                [0.25, -0.25],
-                [0.25, 0.25],
-                [-0.25, -0.25],
-                [-0.25, 0.25],
-                [0.25, 0.75],
-                [0.25, 1.25],
-                [-0.25, 0.75],
-                [-0.25, 1.25],
-                [-0.75, -1.25],
-                [-0.75, -0.75],
-                [-1.25, -1.25],
-                [-1.25, -0.75],
-                [-0.75, -0.25],
-                [-0.75, 0.25],
-                [-1.25, -0.25],
-                [-1.25, 0.25],
-                [-0.75, 0.75],
-                [-0.75, 1.25],
-                [-1.25, 0.75],
-                [-1.25, 1.25],
-            ]
-        )
-    ).all()
-
-    sub_grid_shape = aa.util.grid_2d.grid_2d_slim_via_shape_native_from(
-        shape_native=(2, 4), pixel_scales=(2.0, 1.0), sub_size=3, origin=(0.5, 0.6)
-    )
-
-    sub_grid_mask = aa.util.grid_2d.grid_2d_slim_via_mask_from(
-        mask_2d=np.full(fill_value=False, shape=(2, 4)),
-        pixel_scales=(2.0, 1.0),
-        sub_size=3,
-        origin=(0.5, 0.6),
-    )
-
-    assert (sub_grid_shape == sub_grid_mask).all()
-
 
 def test__grid_2d_via_shape_native_from():
     grid_2d = aa.util.grid_2d.grid_2d_via_shape_native_from(
-        shape_native=(2, 3), pixel_scales=(1.0, 1.0), sub_size=1
+        shape_native=(2, 3),
+        pixel_scales=(1.0, 1.0),
     )
 
     assert (
@@ -344,7 +164,8 @@ def test__grid_2d_via_shape_native_from():
     ).all()
 
     grid_2d = aa.util.grid_2d.grid_2d_via_shape_native_from(
-        shape_native=(3, 2), pixel_scales=(1.0, 1.0), sub_size=1
+        shape_native=(3, 2),
+        pixel_scales=(1.0, 1.0),
     )
 
     assert (
@@ -359,7 +180,7 @@ def test__grid_2d_via_shape_native_from():
     ).all()
 
     grid_2d = aa.util.grid_2d.grid_2d_via_shape_native_from(
-        shape_native=(3, 2), pixel_scales=(1.0, 1.0), sub_size=1, origin=(3.0, -2.0)
+        shape_native=(3, 2), pixel_scales=(1.0, 1.0), origin=(3.0, -2.0)
     )
 
     assert (
@@ -379,7 +200,6 @@ def test__radial_projected_shape_slim_from():
         extent=np.array([-1.0, 1.0, -1.0, 1.0]),
         centre=(0.0, 0.0),
         pixel_scales=(1.0, 1.0),
-        sub_size=1,
     )
 
     assert shape_slim == 2
@@ -388,7 +208,6 @@ def test__radial_projected_shape_slim_from():
         extent=np.array([-1.0, 3.0, -1.0, 1.0]),
         centre=(0.0, 0.0),
         pixel_scales=(1.0, 1.0),
-        sub_size=1,
     )
 
     assert shape_slim == 4
@@ -397,7 +216,6 @@ def test__radial_projected_shape_slim_from():
         extent=np.array([-1.0, 3.0, -1.0, 1.0]),
         centre=(0.0, 1.0),
         pixel_scales=(1.0, 1.0),
-        sub_size=1,
     )
 
     assert shape_slim == 3
@@ -406,7 +224,6 @@ def test__radial_projected_shape_slim_from():
         extent=np.array([-2.0, 1.0, -1.0, 1.0]),
         centre=(0.0, 1.0),
         pixel_scales=(1.0, 1.0),
-        sub_size=1,
     )
 
     assert shape_slim == 4
@@ -417,7 +234,6 @@ def test__grid_scaled_2d_slim_radial_projected_from():
         extent=np.array([-1.0, 1.0, -1.0, 1.0]),
         centre=(0.0, 0.0),
         pixel_scales=(1.0, 1.0),
-        sub_size=1,
     )
 
     assert (grid_radii == np.array([[0.0, 0.0], [0.0, 1.0]])).all()
@@ -426,7 +242,6 @@ def test__grid_scaled_2d_slim_radial_projected_from():
         extent=np.array([-1.0, 3.0, -1.0, 1.0]),
         centre=(0.0, 0.0),
         pixel_scales=(1.0, 1.0),
-        sub_size=1,
     )
 
     assert (
@@ -437,7 +252,6 @@ def test__grid_scaled_2d_slim_radial_projected_from():
         extent=np.array([-1.0, 3.0, -1.0, 1.0]),
         centre=(0.0, 1.0),
         pixel_scales=(1.0, 1.0),
-        sub_size=1,
     )
 
     assert (grid_radii == np.array([[0.0, 1.0], [0.0, 2.0], [0.0, 3.0]])).all()
@@ -446,7 +260,6 @@ def test__grid_scaled_2d_slim_radial_projected_from():
         extent=np.array([-2.0, 1.0, -1.0, 1.0]),
         centre=(0.0, 1.0),
         pixel_scales=(1.0, 1.0),
-        sub_size=1,
     )
 
     assert (
@@ -457,29 +270,6 @@ def test__grid_scaled_2d_slim_radial_projected_from():
         extent=np.array([-1.0, 1.0, -1.0, 1.0]),
         centre=(0.0, 1.0),
         pixel_scales=(0.1, 0.5),
-        sub_size=1,
-    )
-
-    assert (
-        grid_radii
-        == np.array([[0.0, 1.0], [0.0, 1.5], [0.0, 2.0], [0.0, 2.5], [0.0, 3.0]])
-    ).all()
-
-    grid_radii = aa.util.grid_2d.grid_scaled_2d_slim_radial_projected_from(
-        extent=np.array([-1.0, 1.0, -1.0, 1.0]),
-        centre=(0.0, 0.0),
-        pixel_scales=(1.0, 1.0),
-        sub_size=1,
-        shape_slim=3,
-    )
-
-    assert (grid_radii == np.array([[0.0, 0.0], [0.0, 1.0], [0.0, 2.0]])).all()
-
-    grid_radii = aa.util.grid_2d.grid_scaled_2d_slim_radial_projected_from(
-        extent=np.array([-1.0, 1.0, -1.0, 1.0]),
-        centre=(0.0, 1.0),
-        pixel_scales=(0.1, 1.0),
-        sub_size=2,
     )
 
     assert (
@@ -491,7 +281,6 @@ def test__grid_scaled_2d_slim_radial_projected_from():
         extent=np.array([5.0, 8.0, 99.9, 100.1]),
         centre=(100.0, 7.0),
         pixel_scales=(10.0, 0.25),
-        sub_size=1,
     )
 
     assert (
@@ -515,7 +304,6 @@ def test__grid_scaled_2d_slim_radial_projected_from():
         extent=np.array([-1.0, 1.0, -1.0, 3.0]),
         centre=(0.0, 0.0),
         pixel_scales=(1.0, 1.0),
-        sub_size=1,
     )
 
     assert (
@@ -526,7 +314,6 @@ def test__grid_scaled_2d_slim_radial_projected_from():
         extent=np.array([-1.0, 1.0, -2.0, 1.0]),
         centre=(1.0, 0.0),
         pixel_scales=(1.0, 1.0),
-        sub_size=1,
     )
 
     assert (
@@ -537,7 +324,6 @@ def test__grid_scaled_2d_slim_radial_projected_from():
         extent=np.array([-1.0, 1.0, -1.0, 1.0]),
         centre=(1.0, 0.0),
         pixel_scales=(0.5, 0.1),
-        sub_size=1,
     )
 
     assert (
@@ -546,34 +332,9 @@ def test__grid_scaled_2d_slim_radial_projected_from():
     ).all()
 
     grid_radii = aa.util.grid_2d.grid_scaled_2d_slim_radial_projected_from(
-        extent=np.array([-1.0, 1.0, -1.0, 1.0]),
-        centre=(1.0, 0.0),
-        pixel_scales=(0.5, 0.1),
-        sub_size=2,
-    )
-
-    assert (
-        grid_radii
-        == np.array(
-            [
-                [1.0, 0.0],
-                [1.0, 0.25],
-                [1.0, 0.5],
-                [1.0, 0.75],
-                [1.0, 1.0],
-                [1.0, 1.25],
-                [1.0, 1.5],
-                [1.0, 1.75],
-                [1.0, 2.0],
-            ]
-        )
-    ).all()
-
-    grid_radii = aa.util.grid_2d.grid_scaled_2d_slim_radial_projected_from(
         extent=np.array([99.9, 100.1, -1.0, 3.0]),
         centre=(-1.0, 100.0),
         pixel_scales=(1.5, 10.0),
-        sub_size=1,
     )
 
     assert (grid_radii == np.array([[-1.0, 100.0], [-1.0, 101.5], [-1.0, 103.0]])).all()
@@ -591,7 +352,8 @@ def test__grid_2d_slim_from():
     mask = np.array([[True, True, True], [True, False, True], [True, True, True]])
 
     grid_slim = aa.util.grid_2d.grid_2d_slim_from(
-        grid_2d_native=grid_2d, mask=mask, sub_size=1
+        grid_2d_native=grid_2d,
+        mask=mask,
     )
 
     assert (grid_slim == np.array([[5, 5]])).all()
@@ -613,7 +375,8 @@ def test__grid_2d_slim_from():
     )
 
     grid_slim = aa.util.grid_2d.grid_2d_slim_from(
-        grid_2d_native=grid_2d, mask=mask, sub_size=1
+        grid_2d_native=grid_2d,
+        mask=mask,
     )
 
     assert (
@@ -639,47 +402,11 @@ def test__grid_2d_slim_from():
     )
 
     grid_slim = aa.util.grid_2d.grid_2d_slim_from(
-        grid_2d_native=grid_2d, mask=mask, sub_size=1
+        grid_2d_native=grid_2d,
+        mask=mask,
     )
 
     assert (grid_slim == np.array([[2, 2], [4, 4], [5, 5], [6, 6], [8, 8]])).all()
-
-    sub_grid_2d = np.array(
-        [
-            [[1, 1], [2, 2], [3, 3], [4, 4], [5, 5], [6, 6]],
-            [[7, 7], [8, 8], [9, 9], [10, 10], [11, 11], [12, 12]],
-            [[13, 13], [14, 14], [15, 15], [16, 16], [17, 17], [18, 18]],
-            [[1, 1], [2, 2], [3, 3], [4, 4], [5, 5], [6, 6]],
-            [[7, 7], [8, 8], [9, 9], [10, 10], [11, 11], [12, 12]],
-            [[13, 13], [14, 14], [15, 15], [16, 16], [17, 17], [18, 18]],
-        ]
-    )
-
-    mask = np.array([[True, False, True], [True, False, True], [True, True, False]])
-
-    sub_array_2d = aa.util.grid_2d.grid_2d_slim_from(
-        grid_2d_native=sub_grid_2d, mask=mask, sub_size=2
-    )
-
-    assert (
-        sub_array_2d
-        == np.array(
-            [
-                [3, 3],
-                [4, 4],
-                [9, 9],
-                [10, 10],
-                [15, 15],
-                [16, 16],
-                [3, 3],
-                [4, 4],
-                [11, 11],
-                [12, 12],
-                [17, 17],
-                [18, 18],
-            ]
-        )
-    ).all()
 
 
 def test__grid_2d_native_from():
@@ -688,7 +415,8 @@ def test__grid_2d_native_from():
     mask = np.full(fill_value=False, shape=(2, 2))
 
     grid_2d = aa.util.grid_2d.grid_2d_native_from(
-        grid_2d_slim=grid_slim, mask_2d=mask, sub_size=1
+        grid_2d_slim=grid_slim,
+        mask_2d=mask,
     )
 
     assert (
@@ -715,7 +443,8 @@ def test__grid_2d_native_from():
     )
 
     grid_2d = aa.util.grid_2d.grid_2d_native_from(
-        grid_2d_slim=grid_slim, mask_2d=mask, sub_size=1
+        grid_2d_slim=grid_slim,
+        mask_2d=mask,
     )
 
     assert (
@@ -745,24 +474,6 @@ def test__grid_2d_native_from():
             [4.0, 4.0],
         ]
     )
-
-    mask = np.array([[False, False], [False, True]])
-
-    grid_2d = aa.util.grid_2d.grid_2d_native_from(
-        grid_2d_slim=grid_slim, mask_2d=mask, sub_size=2
-    )
-
-    assert (
-        grid_2d
-        == np.array(
-            [
-                [[1.0, 1.0], [1.0, 1.0], [2.0, 2.0], [2.0, 2.0]],
-                [[1.0, 1.0], [1.0, 1.0], [2.0, 2.0], [2.0, 2.0]],
-                [[3.0, 3.0], [3.0, 3.0], [0.0, 0.0], [0.0, 0.0]],
-                [[3.0, 3.0], [4.0, 4.0], [0.0, 0.0], [0.0, 0.0]],
-            ]
-        )
-    ).all()
 
 
 def test__grid_2d_slim_upscaled_from():
